requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
black==22.6.0
click==8.1.3
colorama==0.4.5
coverage==6.4.4
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
executing==0.9.1
fancycompleter==0.9.1
gunicorn==20.1.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
jsonschema==4.16.0
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pdbpp==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
PyJWT==2.4.0
pyrepl==0.9.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
tzdata==2022.2
uritemplate==4.1.1
uuid==1.30
wcwidth==0.2.5
whitenoise==6.2.0
wmctrl==0.4
